:recycle: Split audio logic from path module
