+++ src/io/mod.rs
@@ -1,5 +1,6 @@
 //! Input and output handlers.
 
+pub mod audio;
 pub mod event;
 pub mod path;
 pub mod terminal;
